requirements.txt
tk
bs4
requests
python-nmap
